+++ fpmsyncd/fpmlink.h
@@ -19,7 +19,7 @@ namespace swss {
 class FpmLink : public Selectable {
 public:
     const int MSG_BATCH_SIZE;
-    FpmLink(int port = FPM_DEFAULT_PORT);
+    FpmLink(unsigned short port = FPM_DEFAULT_PORT);
     virtual ~FpmLink();
 
     /* Wait for connection (blocking) */
